LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/at-container-parsing-expected.txt


FAIL (width) assert_equals: expected "true" but got ""
FAIL (min-width: 0px) assert_equals: expected "true" but got ""
FAIL (max-width: 0px) assert_equals: expected "true" but got ""
FAIL (height) assert_equals: expected "true" but got ""
FAIL (min-height: 0px) assert_equals: expected "true" but got ""
FAIL (max-height: 0px) assert_equals: expected "true" but got ""
FAIL (aspect-ratio) assert_equals: expected "true" but got ""
FAIL (min-aspect-ratio: 1/2) assert_equals: expected "true" but got ""
FAIL (max-aspect-ratio: 1/2) assert_equals: expected "true" but got ""
FAIL (orientation: portrait) assert_equals: expected "true" but got ""
FAIL (inline-size) assert_equals: expected "true" but got ""
FAIL (min-inline-size: 0px) assert_equals: expected "true" but got ""
FAIL (max-inline-size: 0px) assert_equals: expected "true" but got ""
FAIL (block-size) assert_equals: expected "true" but got ""
FAIL (min-block-size: 0px) assert_equals: expected "true" but got ""
FAIL (max-block-size: 0px) assert_equals: expected "true" but got ""
FAIL (width: 100px) assert_equals: expected "true" but got ""
FAIL ((width: 100px)) assert_equals: expected "true" but got ""
FAIL (not (width: 100px)) assert_equals: expected "true" but got ""
FAIL ((width: 100px) and (height: 100px)) assert_equals: expected "true" but got ""
FAIL (((width: 40px) or (width: 50px)) and (height: 100px)) assert_equals: expected "true" but got ""
FAIL ((width: 100px) and ((height: 40px) or (height: 50px))) assert_equals: expected "true" but got ""
FAIL (((width: 40px) and (height: 50px)) or (height: 100px)) assert_equals: expected "true" but got ""
FAIL ((width: 50px) or ((width: 40px) and (height: 50px))) assert_equals: expected "true" but got ""
FAIL ((width: 100px) and (not (height: 100px))) assert_equals: expected "true" but got ""
FAIL (width < 100px) assert_equals: expected "true" but got ""
FAIL (width <= 100px) assert_equals: expected "true" but got ""
FAIL (width = 100px) assert_equals: expected "true" but got ""
FAIL (width > 100px) assert_equals: expected "true" but got ""
FAIL (width >= 100px) assert_equals: expected "true" but got ""
FAIL (100px < width) assert_equals: expected "true" but got ""
FAIL (100px <= width) assert_equals: expected "true" but got ""
FAIL (100px = width) assert_equals: expected "true" but got ""
FAIL (100px > width) assert_equals: expected "true" but got ""
FAIL (100px >= width) assert_equals: expected "true" but got ""
FAIL (100px < width < 200px) assert_equals: expected "true" but got ""
FAIL (100px < width <= 200px) assert_equals: expected "true" but got ""
FAIL (100px <= width < 200px) assert_equals: expected "true" but got ""
FAIL (100px > width > 200px) assert_equals: expected "true" but got ""
FAIL (100px > width >= 200px) assert_equals: expected "true" but got ""
FAIL (100px >= width > 200px) assert_equals: expected "true" but got ""
FAIL (width: calc(10px)) assert_equals: expected "true" but got ""
FAIL (width: calc(10em)) assert_equals: expected "true" but got ""
FAIL (width: calc(10px + 10em)) assert_equals: expected "true" but got ""
FAIL (width < calc(10px + 10em)) assert_equals: expected "true" but got ""
FAIL (width < max(10px, 10em)) assert_equals: expected "true" but got ""
FAIL (calc(10px + 10em) < width) assert_equals: expected "true" but got ""
FAIL (calc(10px + 10em) < width < max(30px, 30em)) assert_equals: expected "true" but got ""
FAIL (width: 100px) and (height: 100px) assert_equals: expected "true" but got ""
FAIL (width: 100px) or (height: 100px) assert_equals: expected "true" but got ""
FAIL not (width: 100px) assert_equals: expected "true" but got ""
PASS foo(width)
PASS size(width)
PASS (asdf)
PASS (resolution > 100dpi)
PASS (resolution: 150dpi)
PASS (resolution: calc(2x))
PASS (color)
PASS (min-color: 1)
PASS (color-index >= 1)
PASS size(grid)
PASS (grid)
PASS (width == 100px)
PASS (100px == width)
PASS (100px = width = 200px)
PASS (100px < width > 200px)
PASS (100px <= width >= 200px)
PASS (100px <= width > 200px)
PASS (100px < width >= 200px)
PASS (100px : width : 200px)
PASS screen
PASS print
PASS not print
PASS only print
PASS screen and (width: 100px)
PASS screen or (width: 100px)
PASS not screen and (width: 100px)
PASS not screen or (width: 100px)
PASS (width: 100px), (height: 100px)
PASS foo (width: 100px)
PASS name not (width <= 500px)
PASS not (width <= 500px)
PASS Container name: foo
PASS Container name:  foo
PASS Container name:  foo
PASS Container name: normal
PASS Container name: Normal
PASS Container name: auto
PASS Container name: Auto
PASS Container name: foo foo
PASS Container name: 1px
PASS Container name: 50gil
PASS Container name: name(foo)
PASS Container name: type(inline-size)
PASS Container name: "foo"
PASS Container name: "inherit"
PASS Container name: inherit
PASS Container name: INITIAL
PASS Container name: Unset
PASS Container name: deFAULT
PASS Container name: none
PASS Container name: None
PASS Container name: and
PASS Container name: or
PASS Container name: not
PASS Container name: And
PASS Container name: oR
PASS Container name: nOt
